requirements.txt
accelerate==1.3.0
datasets==3.3.0
fastexcel==0.12.1
matplotlib==3.10.0
numpy==2.2.3
polars==1.22.0
pre-commit==4.1.0
pyside6==6.8.2.1
ruff==0.9.6
scikit-learn==1.6.1
setfit==1.1.1
torch==2.6.0
transformers==4.48.3
umap-learn==0.5.7